增加了用户管理帖子的功能
1.增加了用户管理帖子的功能
已知问题：
1.在我的帖子界面刷新不会跳转回主页
2.在我的帖子界面底部显示的共xx条是当前系统全部的帖子数，而不是个人用户所发布的帖子数
3.ui排版问题

+++ smart-fitness-api/src/main/java/cn/foldedj/mapper/NewsMapper.java
@@ -18,7 +18,7 @@ public interface NewsMapper {
 
     void update(News news);
 
-    void batchDelete(@Param(value = "ids") List<Long> ids);
+    void batchDelete(@Param(value = "ids") List<Integer> ids);
 
     List<NewsVO> query(NewsQueryDto newsQueryDto);
